Refactored api blueprint introduced a utilities file
