--- app/api/__init__.py
@@ -1,3 +1,0 @@
-from flask.blueprints import Blueprint
-
-bp = Blueprint('api', __name__)
