+++ app/utils.py
@@ -1,0 +1,4 @@
+from flask import Blueprint
+
+
+bp = Blueprint('api', __name__)  # needed to enable versioning of my api
